print detailed error message for ErrTermNotFound
Closes #15.

Unfortunately, it seems that the "terminfo" infrastructure in the
Linux/Unix world is defined only at C ABI level (not at file format
level), and as such is fundamentally incompatible with pure-Go
applications. I don't see a good way to solve this problem in a
cross-platform (or, mainly, cross-Unix) way, so I'm currently falling
back to displaying an error message, asking user to patch the mess
locally for themselves by hand.