keys: supplement ctrlKey() with key()
I'm not really sure why KeyCtrlB needs to be wrapped with ctrlKey() on
my machine, if it already has Ctrl in name; I hoped key() would be
enough here. But it seems to only work in ctrlKey() for me. However,
given the confusion, I currently prefer to err on the "beter safe than
sorry" side, and slap both variants everywhere. To be fixed some day in
future maybe.